initial config

+++ src/users/users.service.ts
@@ -1,8 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserInput } from './dto/create-user.input';
 import { UpdateUserInput } from './dto/update-user.input';
 import * as bcrypt from 'bcrypt';
 import { PrismaService } from '../prisma/prisma.service';
+import { CONFIG_MESSAGES } from 'src/config/config';
 
 @Injectable()
 export class UsersService {
@@ -25,20 +26,42 @@ export class UsersService {
     return this.prisma.user.findMany();
   }
 
-  getUserById(id: number) {
-    return this.prisma.user.findUnique({
+  async getUserById(id: number) {
+    const user = await this.prisma.user.findUnique({
       where: { id },
     });
+
+    if (!user) {
+      throw new NotFoundException(CONFIG_MESSAGES.userNotFound);
+    }
+
+    return user;
   }
 
-  updateUser(id: number, updateUserInput: UpdateUserInput) {
+  async updateUserById(id: number, updateUserInput: UpdateUserInput) {
+    const user = await this.prisma.user.findUnique({
+      where: { id },
+    });
+
+    if (!user) {
+      throw new NotFoundException(CONFIG_MESSAGES.userNotFound);
+    }
+
     return this.prisma.user.update({
       where: { id },
       data: updateUserInput,
     });
   }
 
-  deleteUser(id: number) {
+  async deleteUserById(id: number) {
+    const user = await this.prisma.user.findUnique({
+      where: { id },
+    });
+
+    if (!user) {
+      throw new NotFoundException(CONFIG_MESSAGES.userNotFound);
+    }
+
     return this.prisma.user.delete({
       where: { id },
     });
